feat: Moved --export to a seperate export command
